Update pal from commit: 1ac1f17
* [GFX6-10] - Minor simplifications to ColorBlendState objs.
* Resolve occlusion/pipeline stats query: always write availability data
* Add missing support of IL_OP_SAMPLE_XXX_PO to ILP
* [GFX9/10] Reduce the size of Gfx9MaskRam class
* [GFX9-10] Minor tweak to Gfx9Fmask
* [GFX6-10] Remove unneed m_device ref from Gfx*MsaaState
* GFX6-8] Port some minor DepthStencilState space savings from GFX9 hwl
* [GFX6-10] Remove device pointer from several hw-independent state object classes
* Optimize ACE offload for WaitRegMem
* Pal::Amdpg::Platform::ReQueryDevices() - Exiting this function gracefully instead of segfault
* Fix bug in gfx9 perf experiment legacy SQ counter stop sequence
* Setting clock mode through driver ext should fill in the shader and memory clock freq in RGP asic_info chunk
* Fix Vk.api.device_init.create_instance_device_intentional_alloc_fail test failure
* Initialize the enableTcpBigPageTranslationCoalescing setting to a default value
* [GFX6-10] Increase storage for debug only path AutoBuffer
* [GFX9/10] Remove device pointer member from MetaDataAddrEquation
* Bump version number to 225

+++ src/util/imported/metrohash/src/metrohash64.cpp
@@ -1,27 +1,20 @@
 // metrohash64.cpp
 //
-// The MIT License (MIT)
+// Copyright 2015-2018 J. Andrew Rogers
 //
-// Copyright (c) 2015 J. Andrew Rogers
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
 //
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
+//     http://www.apache.org/licenses/LICENSE-2.0
 //
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 //
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-//
+// Modifications Copyright© 2019 Advanced Micro Devices, Inc. All rights reserved.
 
 #include <string.h>
 #include "platform.h"
@@ -85,7 +78,7 @@ void MetroHash64::Update(const uint8_t * const buffer, const uint64_t length)
     }
 
     // bulk update
-    bytes += (end - ptr);
+    bytes += static_cast<uint64_t>(end - ptr);
     while (ptr <= (end - 32))
     {
         // process directly from the source, bypassing the input buffer
@@ -97,7 +90,7 @@ void MetroHash64::Update(const uint8_t * const buffer, const uint64_t length)
 
     // store remaining bytes in input buffer
     if (ptr < end)
-        memcpy(input.b, ptr, end - ptr);
+        memcpy(input.b, ptr, static_cast<size_t>(end - ptr));
 }
